+++ R/fitstats.R
@@ -1358,24 +1358,17 @@ degrees_freedom = function(x, type, vars = NULL, vcov = NULL, se = NULL, cluster
 
     if(is.null(vcov)){
         dof.K = x$nparams
-        t.df = NULL
+        df.t = nobs(x) - dof.K
     } else {
-        t.df = attr(vcov, "G")
+        # From v0.10.0 onward, "dt.t" is always provided!
+        df.t = attr(vcov, "df.t")
         dof.K = attr(vcov, "dof.K")
     }
 
-
-    if(type == "k"){
-        res = dof.K
-    } else if(type == "resid"){
-        res = x$nobs - dof.K
-    } else if(type == "t"){
-        if(is.null(t.df)){
-            res = nobs(x) - dof.K
-        } else {
-            res = t.df - 1
-        }
-    }
+    res = switch(type,
+                 "k" = dof.K,
+                 "resid" = max(nobs(x) - dof.K, 1),
+                 "t" = df.t)
 
     res
 }